+++ core/src/main/java/io/snabble/sdk/encodedcodes/EncodedCodesOptions.java
@@ -13,11 +13,13 @@ public class EncodedCodesOptions {
     public final String nextCodeWithCheck;
     public final boolean repeatCodes;
     public final String countSeparator;
+    public final int maxSizeMm;
     public final Project project;
 
     private EncodedCodesOptions(String prefix, String separator, String suffix, int maxChars,
                                 int maxCodes, String finalCode, String nextCode,
                                 String nextCodeWithCheck, boolean repeatCodes, String countSeparator,
+                                int maxSizeMm,
                                 Project project) {
         this.prefix = prefix;
         this.separator = separator;
@@ -29,6 +31,7 @@ public class EncodedCodesOptions {
         this.nextCodeWithCheck = nextCodeWithCheck;
         this.repeatCodes = repeatCodes;
         this.countSeparator = countSeparator;
+        this.maxSizeMm = maxSizeMm;
         this.project = project;
     }
 
@@ -44,6 +47,7 @@ public class EncodedCodesOptions {
         private int maxCodes = 100;
         private boolean repeatCodes = true;
         private String countSeparator = ";";
+        private int maxSizeMm;
 
         public Builder(Project project) {
             this.project = project;
@@ -99,9 +103,14 @@ public class EncodedCodesOptions {
             return this;
         }
 
+        public Builder maxSizeMm(int maxSizeMm) {
+            this.maxSizeMm = maxSizeMm;
+            return this;
+        }
+
         public EncodedCodesOptions build() {
             return new EncodedCodesOptions(prefix, separator, suffix, maxChars, maxCodes,
-                    finalCode, nextCode, nextCodeWithCheck, repeatCodes, countSeparator, project);
+                    finalCode, nextCode, nextCodeWithCheck, repeatCodes, countSeparator, maxSizeMm, project);
         }
     }
 }