+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [0.13.6]
 
+### Added
+- Support for maximum size of encoded code qr code display
+
 ### Fixed
 - Race condition when updating product database and shopping cart prices at the same time